# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoreBanking API
  version: 1.0.0-phase1
  description: |
    Staff and integration API for CoreBanking. Every request is tenant-scoped: the tenant comes from the
    verified access token (`tenant` claim), never from the URL or body.

    Conventions
    - Money is a decimal **string** (e.g. "100000.00"). Never a JSON number.
    - Dates are ISO `YYYY-MM-DD` business dates; timestamps are RFC 3339 UTC.
    - Changes to master data and money go through maker-checker: the call returns **202** with an
      approval request; the change is applied when a different user approves it.
    - Errors use RFC 9457 `application/problem+json`.
    - `Idempotency-Key` header is honoured on all POSTs that create things.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearer: []
...
paths:
...
  /api/v1/integrations/simulator/callbacks:
    post:
      tags: [Integrations]
      operationId: simulateProviderCallback
      summary: Test and demo only - play a callback of the SIMULATOR provider through the real verification path
      description: 'Permission: integration:simulate.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulatedCallback' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /hooks/v1/{tenant}/{kind}/{provider}:
    post:
      tags: [Integrations]
      operationId: receiveProviderCallback
      security: []
      summary: Provider callback (inbound webhook)
      description: 'Permission: none. No access token - the request is authenticated by the provider signature over the raw body, checked with the secret of the tenant in the path. 200 is returned only after the event is durably stored; a repeated event id is answered 200 with duplicate = true and is not processed again. 401 when the signature does not verify.'
      parameters:
        - { name: tenant, in: path, required: true, schema: { type: string } }
        - { name: kind, in: path, required: true, schema: { type: string, enum: [payout, collection, mandate] } }
        - { name: provider, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/loans/{id}/mandates:
    post:
      tags: [Integrations]
      operationId: registerMandate
      summary: Register an e-mandate for a loan
      description: 'Permission: mandate:register.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MandateInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
    get:
      tags: [Integrations]
      operationId: listLoanMandates
      summary: Mandates of a loan, the one in force first
      description: 'Permission: mandate:view.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/nach-presentations:
    get:
      tags: [Integrations]
      operationId: listLoanNachPresentations
      summary: NACH presentations of a loan with their outcomes
      description: 'Permission: mandate:view.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/mandates:
    get:
      tags: [Integrations]
      operationId: searchMandates
      summary: The mandate register
      description: 'Permission: mandate:view.'
      parameters:
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/mandates/{id}:
    get:
      tags: [Integrations]
      operationId: getMandate
      summary: A mandate with its status history
      description: 'Permission: mandate:view.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/mandates/{id}/status:
    post:
      tags: [Integrations]
      operationId: updateMandateStatus
      summary: Set a mandate status by hand (sponsor bank report, cancellation, suspension)
      description: 'Permission: mandate:admin.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MandateStatusUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/mandates/status-upload:
    post:
      tags: [Integrations]
      operationId: uploadMandateStatuses
      summary: Registration statuses from a file (CSV with mandate_ref, status and optionally umrn, reject_code, reject_reason)
      description: 'Permission: mandate:admin.'
      requestBody:
        required: true
        content:
          text/csv:
            schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/return-reasons:
    get:
      tags: [Integrations]
      operationId: listNachReturnReasons
      summary: NACH return reason codes (indicative until verified)
      description: 'Permission: mandate:view.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/files:
    get:
      tags: [Integrations]
      operationId: listNachFiles
      summary: Presentation and response files
      description: 'Permission: nach:admin.'
      parameters:
        - { name: direction, in: query, description: 'PRESENTATION or RESPONSE', schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/files/{id}/content:
    get:
      tags: [Integrations]
      operationId: getNachFileContent
      summary: The file itself (it holds account numbers; the download is audited)
      description: 'Permission: nach:file.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema: { type: string }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/files/{id}/simulate-response:
    post:
      tags: [Integrations]
      operationId: simulateNachResponse
      summary: Test and demo only - the response a bank would send for a presentation file
      description: 'Permission: integration:simulate.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/presentations/generate:
    post:
      tags: [Integrations]
      operationId: generateNachPresentations
      summary: Generate presentation files for the open business date (the end-of-day step does the same)
      description: 'Permission: nach:admin.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/presentations/pending:
    get:
      tags: [Integrations]
      operationId: listPendingNachFollowUps
      summary: Outcomes whose repayment or bounce charge is not yet posted
      description: 'Permission: nach:admin.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/nach/responses:
    post:
      tags: [Integrations]
      operationId: uploadNachResponse
      summary: Upload a NACH response file for processing
      description: 'Permission: nach:admin. 409 when the same file was already received. The file is refused as a whole when its control totals do not match its rows.'
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/message-templates:
    get:
      tags: [Integrations]
      operationId: listMessageTemplates
      summary: SMS and e-mail templates
      description: 'Permission: message:view.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Integrations]
      operationId: proposeMessageTemplate
      summary: Propose a template (maker-checker)
      description: 'Permission: message:admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MessageTemplateInput' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/message-templates/variables:
    get:
      tags: [Integrations]
      operationId: listMessageTemplateVariables
      summary: Template codes (events) and the placeholders each may use
      description: 'Permission: message:view.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/messages:
    get:
      tags: [Integrations]
      operationId: listMessages
      summary: The delivery log - recipient masked, never the text
      description: 'Permission: message:view.'
      parameters:
        - { name: customerId, in: query, schema: { type: string, format: uuid } }
        - { name: loanId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, description: 'QUEUED, SENT, FAILED or SUPPRESSED', schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/customers/{id}/message-opt-outs:
    post:
      tags: [Integrations]
      operationId: setMessageOptOut
      summary: Record or lift a customer opt-out of a channel
      description: 'Permission: message:admin.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MessageOptOut' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Accepted:
      description: Change recorded as an approval request; applied when a checker approves it
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Approval' }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Money:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,4})?$'
      examples: ['100000.00']
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Field-level problems (validation), when there are several
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, WITHDRAWN]
    Approval:
      type: object
      required: [id, entityType, action, status, maker, madeAt]
      properties:
        id: { type: string, format: uuid }
        entityType: { type: string, examples: [CUSTOMER, BRANCH, BRANCH_SET, STAFF, GL_HEAD, VOUCHER, TAX_RATE, HOLIDAY, TERRITORY, ENUMERATION, SYSTEM_PROPERTY, EOD_SCHEDULE, LOAN_PRODUCT, LOAN_DISBURSEMENT, LOAN_WAIVER, LOAN_REVERSAL, LOAN_AMENDMENT, LOAN_RESTRUCTURE] }
        entityId: { type: [string, 'null'] }
        action: { type: string, examples: [CREATE, UPDATE, REVERSE] }
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        maker: { type: string }
        madeAt: { type: string, format: date-time }
        checker: { type: [string, 'null'] }
        checkedAt: { type: [string, 'null'], format: date-time }
        note: { type: [string, 'null'] }
        amount: { oneOf: [ { $ref: '#/components/schemas/Money' }, { type: 'null' } ] }
        ageHours: { type: number, description: SLA ageing }
        checkersRequired: { type: integer, description: 'Distinct checkers needed (0-3), from the approval rules' }
        approvalsSoFar: { type: integer, description: Approvals already recorded while PENDING }
        appliedRef: { type: [string, 'null'], description: 'What the approval produced (customer id, voucher number …)' }
        current: { type: [object, 'null'], description: State before the change (null for CREATE) }
        proposed: { type: object, description: Proposed state }
...
    IntegrationRecord:
      type: object
      description: A record of the integration module. Money is a decimal string; account numbers appear only masked.
      additionalProperties: true
...
    SimulatedCallback:
      type: object
      required: [kind, reference, status]
      properties:
        kind: { type: string, enum: [payout, collection, mandate] }
        reference: { type: string, description: 'Payout reference, order reference or mandate reference' }
        status: { type: string, description: 'Payout - SUCCESS, FAILED, RETURNED; collection - PAID, FAILED; mandate - ACTIVE, REJECTED, CANCELLED' }
        amount: { $ref: '#/components/schemas/Money' }
        reasonCode: { type: string }
        reason: { type: string }
        eventId: { type: string, description: 'Repeat an id to test replay protection' }
...
    MandateInput:
      type: object
      required: [holderName, accountNumber, ifsc, maxAmount, startDate]
      properties:
        holderName: { type: string, maxLength: 100 }
        accountNumber: { type: string, pattern: '^[A-Za-z0-9]{6,35}$' }
        ifsc: { type: string, pattern: '^[A-Z]{4}0[A-Z0-9]{6}$' }
        accountType: { type: string, enum: [SB, CA, CC, OT], default: SB }
        maxAmount: { $ref: '#/components/schemas/Money' }
        frequency: { type: string, enum: [MONTHLY, QUARTERLY, HALF_YEARLY, YEARLY, AS_PRESENTED], default: MONTHLY }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        sponsorBankCode: { type: string }
        utilityCode: { type: string }
    MandateStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [SUBMITTED, ACTIVE, REJECTED, SUSPENDED, CANCELLED, EXPIRED] }
        umrn: { type: string, pattern: '^[A-Z0-9]{20}$' }
        rejectCode: { type: string }
        rejectReason: { type: string }
...
    MessageTemplateInput:
      type: object
      required: [code, channel, body]
      properties:
        code: { type: string, enum: [LOAN_DISBURSED, PAYMENT_RECEIVED, PAYMENT_BOUNCED, DUE_REMINDER, NOC_ISSUED, RATE_RESET] }
        channel: { type: string, enum: [SMS, EMAIL] }
        language: { type: string, default: en }
        category: { type: string, enum: [TRANSACTIONAL, SERVICE, PROMOTIONAL], default: TRANSACTIONAL }
        subject: { type: string, description: 'E-mail only' }
        body: { type: string, description: 'Text with {{placeholders}}' }
        dltEntityId: { type: string, description: 'SMS - principal entity id on the DLT platform' }
        dltTemplateId: { type: string, description: 'SMS - registered content template id' }
        dltHeader: { type: string, description: 'SMS - registered header (sender id)' }
        status: { type: string, enum: [ACTIVE, RETIRED], default: ACTIVE }
    MessageOptOut:
      type: object
      required: [channel]
      properties:
        channel: { type: string, enum: [SMS, EMAIL] }
        optOut: { type: boolean, default: true }
        source: { type: string }
    # ---- P2-7 platform completion ----
...